Split shared contract domains
