+++ packages/shared/src/workspace.ts
@@ -1,0 +1,20 @@
+export interface SkillImportSelection {
+  token: string
+  directory: string
+}
+
+export interface SandboxStorageStats {
+  root: string
+  totalBytes: number
+  workspaceCount: number
+  referencedWorkspaceCount: number
+  unreferencedWorkspaceCount: number
+  staleWorkspaceCount: number
+  staleThresholdDays: number
+}
+
+export interface SandboxCleanupResult {
+  mode: 'old-unreferenced' | 'all-unreferenced'
+  removedWorkspaces: number
+  removedBytes: number
+}
